requirements.txt
boto3~=1.17.101
PyYAML~=5.4.1
Flask~=2.0.1
marshmallow~=3.12.1